`Task:` Moving components to app folder
