+++ apps/frontend/src/app/LogoSVG/LogoSVG.module.css
