+++ apps/frontend/src/app/LogoSVG/index.tsx
