+++ apps/frontend/src/app/layout.tsx
@@ -3,8 +3,8 @@ import '../styles/app.css'
 import { ReactElement, Suspense } from 'react'
 
 import Footer from '@app/Footer'
-import Grid from '@components/Grid'
-import GridWrap from '@components/GridWrap'
+import Grid from '@app/Grid'
+import GridWrap from '@app/GridWrap'
 import Header from '@app/Header'
 import { Inter } from '@next/font/google'
 import Loading from './loading'
