+++ apps/frontend/src/app/skills/page.tsx
@@ -1,4 +1,4 @@
-import ProjectList from '@components/ProjectItem'
+import ProjectList from '@app/ProjectItem'
 import { ProjectType } from '@lib/sanity.types'
 import { client } from '@lib/sanity.client'
 import { getProjectsData } from '@lib/sanity.queries'
